In preparation to adding the google feast lambdas, rename the apple pubsub and updatesubs files
