+++ typescript/src/feast/update-subs/apple.ts
