+++ webpack.config.js
@@ -20,8 +20,8 @@ const getEntries = (env) => {
     "export-subscription-events-table": "./typescript/src/export/exportEvents.ts",
     "export-historical-data": "./typescript/src/export/exportHistoricalData.ts",
     "apple-revalidate-receipts": "./typescript/src/revalidate-receipts/appleRevalidateReceipts.ts",
-    "feast-apple-pubsub": "./typescript/src/feast/pubsub/pubsub.ts",
-    "feast-apple-update-subscriptions": "./typescript/src/feast/update-subs/updatesubs.ts",
+    "feast-apple-pubsub": "./typescript/src/feast/pubsub/apple.ts",
+    "feast-apple-update-subscriptions": "./typescript/src/feast/update-subs/apple.ts",
   };
   return env.production ? entries : {
     ...entries,
